requirements.txt
Django==1.9.7
sorl-thumbnail==12.3
Pillow==3.3.0
